feat: add Scavenger Hunts feature with dedicated components and styling; update Home page to include navigation to scavenger hunts
