+++ frontend/src/ArtefactDetails/ArtefactDetails.jsx
@@ -1,14 +1,6 @@
 import React from 'react';
 import './ArtefactDetails.css';
 
-const relatedArtefacts = [
-  { id: 1, title: 'Artefact 1', img: '/images/Image-courtesy-of-UP-Museums.jpg' },
-  { id: 2, title: 'Artefact 2', img: '/images/art-2.jpg' },
-  { id: 3, title: 'Artefact 3', img: '/images/art-3.jpg' },
-  { id: 4, title: 'Artefact 4', img: '/images/art-4.jpg' },
-  // Add more if needed
-];
-
 const ArtefactDetails = () => {
   return (
     <div className="artefact-details">
@@ -22,60 +14,44 @@ const ArtefactDetails = () => {
         </button>
       </header>
 
-      {/* Museum Background */}
-      <div className="museum-background">
-        {/* Main Content Overlay */}
-        <div className="content-overlay">
-          {/* Left: Artefact Visual Card */}
-          <div className="artefact-visual">
-            <div className="image-card">
-              <img 
-                src="/images/Image-courtesy-of-UP-Museums.jpg" 
-                alt="Mona Lisa" 
-                className="artefact-image"
-              />
-              <div className="artefact-title-box">
-                <p className="artefact-index">#3</p>
-                <h2 className="artefact-name">Mona Lisa</h2>
-                <p className="artefact-subtitle">Leonardo da Vinci</p>
-              </div>
-            </div>
+      {/* Main Content Container */}
+      <main className="artefact-content">
+        {/* Info Section: Artefact Image and Details */}
+        <div className="info-section">
+          <div className="image-wrapper">
+            <img 
+              src="/images/Image-courtesy-of-UP-Museums.jpg" 
+              alt="Mona Lisa" 
+              className="artefact-image"
+            />
           </div>
-
-          {/* Right: Artefact Info & Description */}
-          <div className="artefact-info">
-            <p className="artefact-meta">
-              <strong>Estimated date:</strong> 1503 &ndash; 1506
-            </p>
-            <p className="artefact-meta">
-              <strong>Where to see it:</strong> Louvre Museum (Paris)
-            </p>
+          <div className="text-info">
+            <h1 className="artefact-title">Mona Lisa</h1>
+            <p className="artefact-subtitle">Leonardo da Vinci</p>
+            <p className="artefact-meta"><strong>Estimated date:</strong> 1503 – 1506</p>
+            <p className="artefact-meta"><strong>Location:</strong> Louvre Museum (Paris)</p>
             <div className="artefact-description">
               <p>
-                Experience the world-renowned masterpiece – the Mona Lisa.
-                Renowned for its enigmatic smile and mysterious allure, this painting 
-                continues to fascinate art lovers around the globe. Discover the subtle 
-                details, innovative techniques, and historical significance behind da Vinci’s 
-                iconic work.
+                The Mona Lisa is one of the world’s most enigmatic masterpieces. Known for its subtle smile and innovative techniques, this painting has captured the imagination of art enthusiasts for centuries.
               </p>
             </div>
           </div>
         </div>
 
-        {/* Carousel Directly Under the Overlay */}
-        <div className="artefact-carousel">
-          {relatedArtefacts.map((item) => (
-            <div key={item.id} className="carousel-item">
-              <img 
-                src={item.img} 
-                alt={item.title} 
-                className="carousel-image" 
-              />
-              <p className="carousel-caption">{item.title}</p>
-            </div>
-          ))}
+        {/* Mascot Section */}
+        <div className="mascot-section">
+          <img 
+            src="/images/mascot.png" 
+            alt="Mascot" 
+            className="mascot-image"
+          />
+          <div className="mascot-bubble">
+            <p>
+              Hi there! I'm Masey, your guide. Did you know that this iconic painting holds secrets that inspire art lovers around the world?
+            </p>
+          </div>
         </div>
-      </div>
+      </main>
     </div>
   );
 };
