--- frontend/src/Home/ScavengerHunts/ScavengerHunts.css
@@ -1,77 +1,0 @@
-/* Container for the scavenger hunts section */
-.scavenger-hunts {
-    background: #fff;
-    padding: 40px 20px;
-    text-align: center;
-    border-top: 1px solid #eaeaea;
-    margin-top: 20px;
-  }
-  
-  /* Section title */
-  .scavenger-hunts-title {
-    font-size: 2.2rem;
-    color: #333;
-    margin-bottom: 30px;
-    font-weight: 600;
-  }
-  
-  /* Layout for hunt cards */
-  .scavenger-hunts-list {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px;
-  }
-  
-  /* Individual hunt card styling */
-  .hunt-card {
-    border: 2px solid;
-    border-radius: 10px;
-    width: 280px;
-    padding: 20px;
-    background-color: #fafafa;
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-  }
-  
-  .hunt-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
-  }
-  
-  /* Header of the card */
-  .hunt-header {
-    padding: 10px;
-    border-top-left-radius: 8px;
-    border-top-right-radius: 8px;
-    color: #fff;
-  }
-  
-  .hunt-header h3 {
-    margin: 0;
-    font-size: 1.4rem;
-    font-weight: 600;
-  }
-  
-  /* Description text */
-  .hunt-description {
-    font-size: 1rem;
-    color: #555;
-    margin: 15px 0;
-  }
-  
-  /* Hunt button styling */
-  .hunt-button {
-    border: none;
-    padding: 10px 20px;
-    color: #fff;
-    border-radius: 25px;
-    cursor: pointer;
-    transition: opacity 0.3s ease;
-    font-size: 1rem;
-  }
-  
-  .hunt-button:hover {
-    opacity: 0.9;
-  }
-
