+++ frontend/src/ScavengerHunts/ScavengerHuntsPage.css
@@ -1,0 +1,104 @@
+/* Define CSS variables for consistency and easy customization */
+:root {
+    --red: #ff5a5f;
+    --blue: #1e90ff;
+    --green: #28a745;
+    --yellow: #ffc107;
+    --text-color: #333;
+    --bg-color: #f7f7f7;
+    --card-bg: #ffffff;
+    --font-family: 'Montserrat', sans-serif;
+    --shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    --shadow-hover: 0 8px 16px rgba(0, 0, 0, 0.2);
+  }
+  
+  .scavenger-hunts {
+    font-family: var(--font-family);
+    background: var(--bg-color);
+    min-height: 100vh;
+    padding: 20px;
+    text-align: center;
+  }
+  
+  .scavenger-hunts-title {
+    font-size: 2.5rem;
+    margin-bottom: 30px;
+    color: var(--text-color);
+  }
+  
+  /* Use a responsive grid to display hunt cards */
+  .scavenger-hunts-list {
+    display: grid;
+    gap: 20px;
+    grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
+    max-width: 1200px;
+    margin: 0 auto;
+  }
+  
+  .hunt-card {
+    background: var(--card-bg);
+    border: 2px solid;
+    border-radius: 12px;
+    overflow: hidden;
+    box-shadow: var(--shadow);
+    transition: transform 0.3s ease, box-shadow 0.3s ease;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .hunt-card:hover {
+    transform: translateY(-5px);
+    box-shadow: var(--shadow-hover);
+  }
+  
+  .hunt-header {
+    padding: 16px;
+    color: #fff;
+    text-align: center;
+  }
+  
+  .hunt-header h3 {
+    margin: 0;
+    font-size: 1.8rem;
+  }
+  
+  .hunt-description {
+    padding: 16px;
+    color: var(--text-color);
+    font-size: 1rem;
+    flex: 1;
+    text-align: left;
+  }
+  
+  .hunt-button {
+    border: none;
+    color: #fff;
+    padding: 12px 24px;
+    font-size: 1rem;
+    border-radius: 25px;
+    margin: 16px;
+    cursor: pointer;
+    transition: background 0.3s ease, transform 0.3s ease;
+  }
+  
+  .hunt-button:hover {
+    transform: scale(1.03);
+  }
+  
+  /* Responsive adjustments for smaller screens */
+  @media (max-width: 480px) {
+    .scavenger-hunts-title {
+      font-size: 2rem;
+    }
+    .hunt-header h3 {
+      font-size: 1.5rem;
+    }
+    .hunt-description {
+      font-size: 0.95rem;
+    }
+    .hunt-button {
+      font-size: 0.9rem;
+      padding: 10px 18px;
+    }
+  }
+
