+++ frontend/src/ScavengerHunts/ScavengerHuntsPage.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import "./ScavengerHunts.css";
+import  ScavengerHunts from './ScavengerHunts';
+
+
+const ScavengerHuntsPage = () => {
+
+  return (
+<ScavengerHunts></ScavengerHunts>
+  );
+};
+
+export default ScavengerHuntsPage;
